+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Documentation in English and German ([#6](https://github.com/scm-manager/scm-ssh-plugin/pull/6))
+### Fixed
+- Anonymous users should not be allowed to access or modify authorized keys ([#9](https://github.com/scm-manager/scm-ssh-plugin/pull/9))
 
 ## 2.0.0 - 2020-06-04
 ### Changed